refactoring. client: add login.
