+++ RawCMSClient/RawCMSClient.csproj
@@ -7,6 +7,8 @@
 
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.3.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
+    <PackageReference Include="RestSharp" Version="106.5.4" />
   </ItemGroup>
 
 </Project>
